requirements.txt

opencv-python
numpy<2
insightface
onnxruntime-gpu
torch
torchvision
mediapipe
Pillow